fix: Enhance the exectuion logic, input/output redirection and error resolution in the pipeline.
Description:
This commit resolves the issue related to arguments, specifically the usage of arguments like [/dev/random "cat" "head -10" outfile], and ensures that child processes are executed and synchronized correctly.

Changes Made:
- parent_process(int infile, int *fd): This function handles the
cleanups of the parent process. It closes the input file descriptor,
duplicates the read end of the pipe into the input file descriptor, and
then closes both ends of the pipe.
- child_process(t_cmd **cmd, int *fd, int idx): The child process
execution logic is organized into this function. It closes both ends of
the pipe and then employs 'execve()' to execute the specified command.
If execution fails, the child process exits with an error code.
- create_process(t_cmd *cmd, t_file file, int cnt): This function
orchestrates the creation and execution of child processes for each
command in the pipeline. It manages input/output redirection and
utilizes 'dup2()' to set up proper file descriptors for each process.
The 'waitpid()' function waits for child processes to complete.
- The nesting issue in the previous version was fixed. The
'waitpid()' loop is now properly positioned outside the loop that
creates child processes.

Note:
The bonus part of the project is yet to be completed.